Added app testsuite to activate memory stress
Stressful Application Test (or stressapptest, its unix name) is
a memory interface test. It tries to maximize randomized traffic
to memory from processor and I/O, with the intent of creating
a realistic high load situation in order to test the existing
hardware devices in a computer. It has been used at Google for
some time and now it is available under the apache 2.0 license.

(Exported from code.google.com/p/stressapptest)

Signed-off-by: Geunsik Lim <leemgs@gmail.com>